+++ doc/changes.md
@@ -409,6 +409,7 @@ as part of K95 at this time, the default terminal remains VT220 for now.
  - DSR-85 Multiple Session Configuration Status Report (VT420 and up) now 
    reports not configured.
  - [ED-3 (xt)](https://davidrg.github.io/ckwin/dev/ctlseqs.html#ed-3xt) - Clear Scrollback (K95 only), xterm extension
+ - [DECST8C](https://davidrg.github.io/ckwin/dev/ctlseqs.html#decst8c) - Set Tab at Every 8 Columns (VT520)
 
 ### Fixed Bugs
  - Fixed an issue introduced in beta 7 which could cause SSH connections made